# internal/server/web/openapi.yaml
...
openapi: 3.1.0
info:
  title: wudict
  version: "1"
  summary: A local dictionary server for MDX/MDD, StarDict, Slob, DSL and BGL files.
  description: |
    WuWeiDict serves your own dictionary files over HTTP, on your own machine.

    **Streaming.** `/api/dicts`, `/api/search` and `/api/rescan` answer with
    NDJSON: one JSON object per line, flushed the moment it is ready. Read them
    line by line; do not wait for the body to end. That is the whole point - the
    first dictionary's results arrive while the slowest one is still opening.
    Every stream opens with a `begin` line and closes with an `end` line.

    `/api/ingest` streams Server-Sent Events instead, because it reports
    progress rather than results.

    **Switches.** Every boolean query parameter reads the same spellings: off
    for `0`, `false`, `no`, `off` or an empty value, on for anything else.

    **Two surfaces.** Three read-only endpoints answer a browser extension
    cross-origin: `/api/dicts`, `/api/search` and `/res/`. Everything else is
    same-origin, is marked `internal` below, and may change between releases.

    **An access key, when the server is reachable.** On the default
    `127.0.0.1` bind there is nothing to authenticate and no key is asked for.
    On any other address - and Go's wildcard listen is dual-stack, so `0.0.0.0`
    includes every globally routable IPv6 address the machine holds - `AUTH`
...
    and let the browser keep the cookie it sets. Without it: `401`, with
...

...

...
servers:
...
security: []
...
paths:
...
  /api/groups:
    get:
      tags: [internal]
      operationId: getGroups
      x-internal: true
      summary: The groups the dictionary picker offers
      description: >
        `groups.ini`, beside the `wudict.toml` in effect: a `[Section]` line is
        a "Group by" heading (at most 32, names of at most 64 bytes), and
        `Name = text, text` a group under it that a dictionary joins when one
        of the texts (at least 2 characters) is anywhere in its title or file
        name, in any case. An
        item between backticks is a regular expression (Go RE2,
        case-insensitive), read verbatim and tested against the title and the
        file name separately. `;` and `#` start a comment outside backticks. A
        group line before any section belongs to `[My groups]`. While there is no file the built-in list is in effect
        (`custom` false); once saved, the file is the whole list.


        Ids are the names, trimmed and lower-cased: a dictionary row's
        `groups[].f` and `groups[].v` (`GET /api/dicts`) are exactly these.
        `problems` lists the lines and items that were skipped (an invalid
        regular expression, a line with no `=`, a one-character item, more
        than 500 regular expressions); the rest of the file still applies. At
        most 100 are listed: a last entry with `more` stands for the rest.
        `text` and `msg` are clipped to 200 bytes. `unusable` means the file exists but could not be read: the
        built-in list is in effect, and `problems` has one entry, with no line,
        saying why.
      x-codeSamples:
        - lang: Shell
          label: curl
          source: |
            curl -s 'http://127.0.0.1:6888/api/groups' | jq -r .text
      responses:
        '200':
          description: The file in effect.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Groups' }
    put:
      tags: [internal]
      operationId: saveGroups
      x-internal: true
      summary: Replace groups.ini
      description: >
        The body is the whole file, as text. Lines that do not parse are
        reported in `problems`, never refused. A body identical to the built-in
        list removes the file instead, so an unchanged copy does not stop the
        list from following the built-in one. Capped at 256 KiB. While the
        current file is `unusable`, a save is refused unless `replace=1` is
        given: the file is the user's, and the editor could not show it.
      parameters:
        - name: replace
          in: query
          required: false
          description: 1 overwrites a groups.ini that exists but cannot be read.
          schema: { type: string, enum: ['1'] }
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
      responses:
        '200':
          description: The file in effect after the save.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Groups' }
        '400':
          description: The body is not UTF-8 text, or could not be read.
        '409':
          description: >
            There is no config directory, so there is nowhere to save; or the
            current file cannot be read and `replace=1` was not given.
        '413':
          description: The body is larger than 256 KiB.
        '500':
          description: The file could not be written.
    delete:
      tags: [internal]
      operationId: resetGroups
      x-internal: true
      summary: Back to the built-in groups
      description: Removes groups.ini; the built-in list is in effect again.
      responses:
        '200':
          description: The built-in list, now in effect.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Groups' }
        '500':
          description: The file could not be removed.
...
components:
...
  schemas:
...
    Groups:
      type: object
      description: The groups.ini in effect, what it defines, and the lines skipped.
      required: [text, custom, unusable, file, writable, problems, facets]
      properties:
        text:
          type: string
          description: The file's text; the built-in list when there is no file, empty when the file cannot be read.
        custom:
          type: boolean
          description: A groups.ini exists. False means the built-in list is in effect.
        unusable:
          type: boolean
          description: The groups.ini that exists cannot be read; the built-in list is in effect.
        file:
          type: string
          description: Where groups.ini is (or would be) kept; empty when there is no config directory.
        writable:
          type: boolean
          description: False when there is no config directory, so nothing can be saved.
        problems:
          type: array
          items:
            type: object
            properties:
              line: { type: integer, description: 1-based; 0 when the file as a whole could not be used. }
              text: { type: string, description: The line as written. }
              msg: { type: string, description: Why it was skipped. }
              more: { type: integer, description: 'On a last entry only: how many problems it stands for, not listed.' }
        facets:
          type: array
          description: The sections and their groups, in file order.
          items:
            type: object
            properties:
              id: { type: string }
              label: { type: string }
              groups:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string }
                    label: { type: string }
                    items:
                      type: array
                      items: { type: string }
                      description: What the group matches, as parsed - plain items lower-cased, regular expressions in backticks.
...